适配最新合并。怼老的MessageHandler的适配。可以对老代码无需更改自动适配
