+++ example/src/main/assets/demo.html
@@ -22,12 +22,13 @@
     <input type="text" id="text2" value="password"/>
 </p>
 <p>
-    <input type="button" id="enter" value="发消息给Native" onclick="testClick();"
-    />
+    <input type="button" id="enter" value="发消息给Native" onclick="testClick();"/>
 </p>
 <p>
-    <input type="button" id="enter1" value="调用Native方法" onclick="testClick1();"
-    />
+    <input type="button" id="enter1" value="调用Native方法" onclick="testClick1();"/>
+</p>
+<p>
+    <input type="button" id="enter3" value="调用Native方法2" onclick="testClick2();"/>
 </p>
 <p>
     <input type="button" id="enter2" value="显示html" onclick="testDiv();"/>
@@ -50,7 +51,7 @@
             window.WebViewJavascriptBridge.send(
                 data
                 , function(responseData) {
-                    document.getElementById("show").innerHTML = "repsonseData from java, data = " + responseData
+                    document.getElementById("show").innerHTML = "response Data from java, data = " + responseData
                 }
             );
 
@@ -63,6 +64,20 @@
             //call native method
             window.WebViewJavascriptBridge.callHandler(
                 'submitFromWeb'
+                , {'param': '中文测试'}
+                , function(responseData) {
+                    document.getElementById("show").innerHTML = "send get responseData from java, data = " + responseData
+                }
+            );
+        }
+
+        function testClick2() {
+            var str1 = document.getElementById("text1").value;
+            var str2 = document.getElementById("text2").value;
+
+            //call native method
+            window.WebViewJavascriptBridge.callHandler(
+                'submitFromWeb2'
                 , {'param': '中文测试'}
                 , function(responseData) {
                     document.getElementById("show").innerHTML = "send get responseData from java, data = " + responseData
@@ -109,6 +124,8 @@
                 }
             });
         })
+
+
 
 </script>
