+++ library/src/main/java/com/github/lzyzsd/jsbridge/BridgeHandler.java
@@ -1,0 +1,8 @@
+package com.github.lzyzsd.jsbridge;
+
+@Deprecated
+public interface BridgeHandler {
+	
+	void handler(String data, CallBackFunction function);
+
+}
