+++ library/src/main/java/com/github/lzyzsd/jsbridge/BridgeHelper.java
@@ -1,15 +1,18 @@
 package com.github.lzyzsd.jsbridge;
 
+import android.os.Build;
 import android.os.Looper;
 import android.os.SystemClock;
 import android.text.TextUtils;
-import android.util.Log;
+import android.webkit.ValueCallback;
 
-import java.io.UnsupportedEncodingException;
-import java.net.URLDecoder;
+import androidx.annotation.RequiresApi;
+import androidx.collection.ArrayMap;
+
+import com.google.gson.Gson;
+
 import java.net.URLEncoder;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -19,109 +22,231 @@ import java.util.Map;
  * @author ZhengAn
  * @date 2019-06-30
  */
-public class BridgeHelper implements WebViewJavascriptBridge {
+public class BridgeHelper implements WebViewJavascriptBridge, WebViewClientProxy.OnLoadJSListener {
 
-    private static final String TAG = "BridgeHelper";
+    @Deprecated
+    private Map<String, BridgeHandler> mMessageHandlers = new ArrayMap<>();
+    private Map<String, CallBackFunction> mCallbacks = new ArrayMap<>();
+    @Deprecated
+    private BridgeHandler mDefaultHandler = new DefaultHandler();
 
-    private static final String BRIDGE_JS = "WebViewJavascriptBridge.js";
-    private Map<String, CallBackFunction> responseCallbacks = new HashMap<>();
-    private Map<String, BridgeHandler> messageHandlers = new HashMap<>();
-    private BridgeHandler defaultHandler = new DefaultHandler();
+    private List<Object> mMessages = new ArrayList<>();
 
-    private List<Message> startupMessage = new ArrayList<>();
+    private long mUniqueId = 0;
 
-    private List<Message> getStartupMessage() {
-        return startupMessage;
-    }
+    private boolean mJSLoaded = false;
 
-    private void setStartupMessage(List<Message> startupMessage) {
-        this.startupMessage = startupMessage;
-    }
-
-    private long uniqueId = 0;
+    private Gson mGson = new Gson();
 
     private IWebView webView;
 
     public BridgeHelper(IWebView webView) {
         this.webView = webView;
+        webView.addJavascriptInterface(new BaseJavascriptInterface(this), "android");
+        BridgeWebViewClient client = new BridgeWebViewClient() {
+
+            @Override
+            public void onBridgeLoadStart() {
+                mJSLoaded = false;
+            }
+
+            @Override
+            public void onBridgeLoadFinished() {
+                mJSLoaded = true;
+                if (mMessages != null) {
+                    for (Object message : mMessages) {
+                        dispatchMessage(message);
+                    }
+                    mMessages = null;
+                }
+            }
+        };
+        webView.setWebViewClient(client);
+    }
+
+    public void setGson(Gson gson) {
+        mGson = gson;
+    }
+
+    public boolean isJSLoaded() {
+        return mJSLoaded;
+    }
+
+    final Map<String, CallBackFunction> getResponseCallbacks() {
+        return mCallbacks;
+    }
+
+    final Map<String, BridgeHandler> getMessageHandlers() {
+        return mMessageHandlers;
+    }
+
+    final BridgeHandler getDefaultHandler() {
+        return mDefaultHandler;
+    }
+
+    @Override
+    public void onLoadStart() {
+        mJSLoaded = false;
+    }
+
+    @Override
+    public void onLoadFinished() {
+        mJSLoaded = true;
+        if (mMessages != null) {
+            for (Object message : mMessages) {
+                dispatchMessage(message);
+            }
+            mMessages = null;
+        }
+    }
+
+    @Override
+    public void sendToWeb(Object data) {
+        sendToWeb(data, (CallBackFunction) null);
+    }
+
+    @Override
+    public void sendToWeb(Object data, CallBackFunction responseCallback) {
+        doSend(null, data, responseCallback);
+    }
+
+    /**
+     * call javascript registered handler
+     * 调用javascript处理程序注册
+     *
+     * @param handlerName handlerName
+     * @param data        data
+     * @param callBack    OnBridgeCallback
+     */
+    public void callHandler(String handlerName, String data, CallBackFunction callBack) {
+        doSend(handlerName, data, callBack);
     }
 
     /**
      * @param handler default handler,handle messages send by js without assigned handler name,
      *                if js message has handler name, it will be handled by named handlers registered by native
      */
+    @Deprecated
     public void setDefaultHandler(BridgeHandler handler) {
-        this.defaultHandler = handler;
+        this.mDefaultHandler = handler;
     }
 
     /**
-     * 获取到CallBackFunction data执行调用并且从数据集移除
+     * register handler,so that javascript can call it
+     * 注册处理程序,以便javascript调用它
      *
-     * @param url
+     * @param handlerName handlerName
+     * @param handler     BridgeHandler
      */
-    private void handlerReturnData(String url) {
-        String functionName = BridgeUtil.getFunctionFromReturnUrl(url);
-        CallBackFunction f = responseCallbacks.get(functionName);
-        String data = BridgeUtil.getDataFromReturnUrl(url);
-        if (f != null) {
-            f.onCallBack(data);
-            responseCallbacks.remove(functionName);
+    @Deprecated
+    public void registerHandler(String handlerName, BridgeHandler handler) {
+        if (handler != null) {
+            // 添加至 Map<String, BridgeHandler>
+            mMessageHandlers.put(handlerName, handler);
         }
     }
 
-    @Override
-    public void send(String data) {
-        send(data, null);
+    /**
+     * unregister handler
+     *
+     * @param handlerName
+     */
+    @Deprecated
+    public void unregisterHandler(String handlerName) {
+        if (handlerName != null) {
+            mMessageHandlers.remove(handlerName);
+        }
     }
 
+    /**
+     * Experimental. Use with caution
+     *
+     * @param function method name
+     * @param values   method parameters
+     */
     @Override
-    public void send(String data, CallBackFunction responseCallback) {
-        doSend(null, data, responseCallback);
+    public void sendToWeb(String function, Object... values) {
+        // 必须要找主线程才会将数据传递出去 --- 划重点
+        if (Thread.currentThread() == Looper.getMainLooper().getThread()) {
+            String jsCommand = String.format(function, values);
+            jsCommand = String.format(BridgeUtil.JAVASCRIPT_STR, jsCommand);
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+                webView.evaluateJavascript(jsCommand, null);
+            } else {
+                webView.loadUrl(jsCommand);
+            }
+        }
     }
+
+
+    /**
+     * Experimental. Use with caution
+     *
+     * @param function method name
+     * @param callback value callback
+     * @param values   method parameters
+     */
+    @Override
+    @RequiresApi(api = Build.VERSION_CODES.KITKAT)
+    public void sendToWeb(String function, ValueCallback<String> callback, Object... values) {
+        // 必须要找主线程才会将数据传递出去 --- 划重点
+        if (Thread.currentThread() == Looper.getMainLooper().getThread()) {
+            String jsCommand = String.format(function, values);
+            jsCommand = String.format(BridgeUtil.JAVASCRIPT_STR, jsCommand);
+            webView.evaluateJavascript(jsCommand, callback);
+        }
+    }
+
 
     /**
      * 保存message到消息队列
      *
      * @param handlerName      handlerName
      * @param data             data
-     * @param responseCallback CallBackFunction
+     * @param responseCallback OnBridgeCallback
      */
-    private void doSend(String handlerName, String data, CallBackFunction responseCallback) {
-        Message m = new Message();
-        if (!TextUtils.isEmpty(data)) {
-            m.setData(data);
+    private void doSend(String handlerName, Object data, CallBackFunction responseCallback) {
+        if (!(data instanceof String) && mGson == null) {
+            return;
+        }
+        JSRequest request = new JSRequest();
+        if (data != null) {
+            request.data = data instanceof String ? (String) data : mGson.toJson(data);
         }
         if (responseCallback != null) {
-            String callbackStr = String.format(BridgeUtil.CALLBACK_ID_FORMAT, ++uniqueId + (BridgeUtil.UNDERLINE_STR + SystemClock.currentThreadTimeMillis()));
-            responseCallbacks.put(callbackStr, responseCallback);
-            m.setCallbackId(callbackStr);
+            String callbackId = String.format(BridgeUtil.CALLBACK_ID_FORMAT, (++mUniqueId) + (BridgeUtil.UNDERLINE_STR + SystemClock.currentThreadTimeMillis()));
+            mCallbacks.put(callbackId, responseCallback);
+            request.callbackId = callbackId;
         }
         if (!TextUtils.isEmpty(handlerName)) {
-            m.setHandlerName(handlerName);
+            request.handlerName = handlerName;
         }
-        queueMessage(m);
+        queueMessage(request);
     }
 
     /**
      * list<message> != null 添加到消息集合否则分发消息
      *
-     * @param m Message
+     * @param message Message
      */
-    private void queueMessage(Message m) {
-        if (startupMessage != null) {
-            startupMessage.add(m);
+    private void queueMessage(Object message) {
+        if (mMessages != null) {
+            mMessages.add(message);
         } else {
-            dispatchMessage(m);
+            dispatchMessage(message);
         }
     }
 
     /**
      * 分发message 必须在主线程才分发成功
      *
-     * @param m Message
+     * @param message Message
      */
-    private void dispatchMessage(Message m) {
-        String messageJson = m.toJson();
+    private void dispatchMessage(Object message) {
+        if (mGson == null) {
+            return;
+        }
+        String messageJson = mGson.toJson(message);
         //escape special characters for json string  为json字符串转义特殊字符
         messageJson = messageJson.replaceAll("(\\\\)([^utrn])", "\\\\\\\\$1$2");
         messageJson = messageJson.replaceAll("(?<=[^\\\\])(\")", "\\\\\"");
@@ -132,155 +257,40 @@ public class BridgeHelper implements WebViewJavascriptBridge {
         String javascriptCommand = String.format(BridgeUtil.JS_HANDLE_MESSAGE_FROM_JAVA, messageJson);
         // 必须要找主线程才会将数据传递出去 --- 划重点
         if (Thread.currentThread() == Looper.getMainLooper().getThread()) {
-            this.loadUrl(javascriptCommand);
-        }
-    }
-
-    /**
-     * 刷新消息队列
-     */
-    private void flushMessageQueue() {
-        if (Thread.currentThread() == Looper.getMainLooper().getThread()) {
-            loadUrl(BridgeUtil.JS_FETCH_QUEUE_FROM_JAVA, new CallBackFunction() {
-
-                @Override
-                public void onCallBack(String data) {
-                    // deserializeMessage 反序列化消息
-                    List<Message> list = null;
-                    try {
-                        list = Message.toArrayList(data);
-                    } catch (Exception e) {
-                        Log.w(TAG, e);
-                        return;
-                    }
-                    if (list == null || list.isEmpty()) {
-                        return;
-                    }
-                    for (int i = 0; i < list.size(); i++) {
-                        Message m = list.get(i);
-                        String responseId = m.getResponseId();
-                        // 是否是response  CallBackFunction
-                        if (!TextUtils.isEmpty(responseId)) {
-                            CallBackFunction function = responseCallbacks.get(responseId);
-                            String responseData = m.getResponseData();
-                            function.onCallBack(responseData);
-                            responseCallbacks.remove(responseId);
-                        } else {
-                            CallBackFunction responseFunction = null;
-                            // if had callbackId 如果有回调Id
-                            final String callbackId = m.getCallbackId();
-                            if (!TextUtils.isEmpty(callbackId)) {
-                                responseFunction = new CallBackFunction() {
-                                    @Override
-                                    public void onCallBack(String data) {
-                                        Message responseMsg = new Message();
-                                        responseMsg.setResponseId(callbackId);
-                                        responseMsg.setResponseData(data);
-                                        queueMessage(responseMsg);
-                                    }
-                                };
-                            } else {
-                                responseFunction = new CallBackFunction() {
-                                    @Override
-                                    public void onCallBack(String data) {
-                                        // do nothing
-                                    }
-                                };
-                            }
-                            // BridgeHandler执行
-                            BridgeHandler handler;
-                            if (!TextUtils.isEmpty(m.getHandlerName())) {
-                                handler = messageHandlers.get(m.getHandlerName());
-                            } else {
-                                handler = defaultHandler;
-                            }
-                            if (handler != null) {
-                                handler.handler(m.getData(), responseFunction);
-                            }
-                        }
-                    }
-                }
-            });
-        }
-    }
-
-    private void loadUrl(String jsUrl, CallBackFunction returnCallback) {
-        this.loadUrl(jsUrl);
-        // 添加至 Map<String, CallBackFunction>
-        responseCallbacks.put(BridgeUtil.parseFunctionName(jsUrl), returnCallback);
-    }
-
-    private void loadUrl(String jsUrl) {
-        webView.loadUrl(jsUrl);
-    }
-
-    /**
-     * register handler,so that javascript can call it
-     * 注册处理程序,以便javascript调用它
-     *
-     * @param handlerName handlerName
-     * @param handler     BridgeHandler
-     */
-    public void registerHandler(String handlerName, BridgeHandler handler) {
-        if (handler != null) {
-            // 添加至 Map<String, BridgeHandler>
-            messageHandlers.put(handlerName, handler);
-        }
-    }
-
-    /**
-     * unregister handler
-     *
-     * @param handlerName
-     */
-    public void unregisterHandler(String handlerName) {
-        if (handlerName != null) {
-            messageHandlers.remove(handlerName);
-        }
-    }
-
-    /**
-     * call javascript registered handler
-     * 调用javascript处理程序注册
-     *
-     * @param handlerName handlerName
-     * @param data        data
-     * @param callBack    CallBackFunction
-     */
-    public void callHandler(String handlerName, String data, CallBackFunction callBack) {
-        doSend(handlerName, data, callBack);
-    }
-
-    public void onPageFinished() {
-        webViewLoadLocalJs();
-
-        if (getStartupMessage() != null) {
-            for (Message m : getStartupMessage()) {
-                dispatchMessage(m);
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+                webView.evaluateJavascript(javascriptCommand, null);
+            } else {
+                webView.loadUrl(javascriptCommand);
             }
-            setStartupMessage(null);
         }
     }
 
-    private void webViewLoadLocalJs() {
-        String jsContent = BridgeUtil.assetFile2Str(webView.getContext(), BridgeHelper.BRIDGE_JS);
-        loadUrl("javascript:" + jsContent);
+    public void sendResponse(Object data, String callbackId) {
+        if (!(data instanceof String) && mGson == null) {
+            return;
+        }
+        if (!TextUtils.isEmpty(callbackId)) {
+            final JSResponse response = new JSResponse();
+            response.responseId = callbackId;
+            response.responseData = data instanceof String ? (String) data : mGson.toJson(data);
+            if (Thread.currentThread() == Looper.getMainLooper().getThread()) {
+                dispatchMessage(response);
+            } else {
+                webView.post(new Runnable() {
+                    @Override
+                    public void run() {
+                        dispatchMessage(response);
+                    }
+                });
+            }
+
+        }
     }
 
-    public boolean shouldOverrideUrlLoading(String url) {
-        try {
-            url = URLDecoder.decode(url, "UTF-8");
-        } catch (UnsupportedEncodingException e) {
-            Log.w(TAG, e);
-        }
-
-        if (url.startsWith(BridgeUtil.YY_RETURN_DATA)) { // 如果是返回数据
-            handlerReturnData(url);
-            return true;
-        } else if (url.startsWith(BridgeUtil.YY_OVERRIDE_SCHEMA)) { //
-            flushMessageQueue();
-            return true;
-        }
-        return false;
+    public void destroy() {
+        mCallbacks.clear();
+        mMessageHandlers.clear();
     }
+
+
 }
