+++ library/src/main/java/com/github/lzyzsd/jsbridge/CallBackFunction.java
@@ -1,0 +1,11 @@
+package com.github.lzyzsd.jsbridge;
+
+/**
+ * use {@link OnBridgeCallback}
+ */
+@Deprecated
+public interface CallBackFunction {
+	
+	void onCallBack(String data);
+
+}
