+++ library/src/main/java/com/github/lzyzsd/jsbridge/DefaultHandler.java
@@ -1,0 +1,13 @@
+package com.github.lzyzsd.jsbridge;
+
+@Deprecated
+public class DefaultHandler implements BridgeHandler{
+	
+	@Override
+	public void handler(String data, CallBackFunction function) {
+		if(function != null){
+			function.onCallBack("DefaultHandler response data");
+		}
+	}
+
+}
