+++ library/src/main/java/com/github/lzyzsd/jsbridge/IWebView.java
@@ -1,6 +1,9 @@
 package com.github.lzyzsd.jsbridge;
 
-import android.content.Context;
+import android.webkit.ValueCallback;
+import android.webkit.WebViewClient;
+
+import androidx.annotation.Nullable;
 
 /**
  * WebView功能接口.
@@ -10,7 +13,13 @@ import android.content.Context;
  */
 public interface IWebView {
 
-    Context getContext();
+    void addJavascriptInterface(Object obj, String interfaceName);
+
+    void setWebViewClient(WebViewClient client);
+
+    void evaluateJavascript(String script, @Nullable ValueCallback<String> resultCallback);
 
     void loadUrl(String url);
+
+    boolean post(Runnable runnable);
 }
